fix #31: Fix addNumber throwing on valid sudoku
AddNumber was throwing on a valid sudoku because there was a mixup in row
and column arguments passed to the setElement function, should be fixed
now.
Added a test to ensure it throws in the cases where it is supposed to.
Unsure if we should set up a test for the normal case, as that is going to
be covered in most of the other tests

ref #13
ref #10